refactor and create input plugin
